requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-decouple==3.8
loguru==0.7.2
factory-boy==3.3.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
django-ratelimit==4.1.0
django-csp==3.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
Faker==22.0.0